Enhance radio settings and improve environment variable handling
- Added new radio settings for speaker parameters including upper breakup frequency, Q factor, gain, cone dip frequency, Q factor, gain, and various tolerances.
- Refactored environment variable retrieval to use a new helper function `getEnvString` for better readability and safety.
- Updated file opening to support UTF-8 paths with `openFileUtf8`.
- Improved input handling in the TUI by replacing direct getenv calls with the new helper function.
- Cleaned up unused variables and improved type safety in various places, including static casts for clarity.
- Enhanced video processing and playback logic with better error handling and logging.
- Removed unnecessary diagnostics and streamlined fullscreen handling in the video window.

+++ audiofilter/radio1938/analysis/radio_calibration.cpp
@@ -299,8 +299,8 @@ void updateCalibrationSnapshot(Radio1938& radio) {
     radio.calibration.outputGridBPositiveFraction =
         radio.calibration.outputGridBPositiveSamples * invValidationCount;
     radio.calibration.interstageSecondaryRmsVolts =
-        std::sqrt(radio.calibration.interstageSecondarySumSq *
-                  invValidationCount);
+        static_cast<float>(std::sqrt(radio.calibration.interstageSecondarySumSq *
+                                     invValidationCount));
   }
 
   const auto& receiverPass =
